Fix #663 ADIOS Avoid 32bit Down-Cast
`defineAdiosVar` caused a down-cast to 32bit for counter
attributes such as offsets.
That caused problems with counters and offsets for particles,
that are traditionally stored in a long 1D array and can
exceed the size of uint32_t (the base type of `DataSpace<DIM>`)
easily.

First spotted by @psychocoderHPC - thanks!